src/web-admin/backend/requirements.txt
# Web Admin Backend Dependencies

# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# MinIO client
minio==7.2.3

# Qdrant client
qdrant-client==1.7.0

# Embeddings
sentence-transformers==2.2.2
torch==2.1.0

# File processing
pandas==2.1.4
openpyxl==3.1.2
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.1.0

# Configuration
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.6

# HTTP client
httpx==0.28.1
aiofiles==23.2.1
